Address review feedback for AwsStsStrategy
- Rename TokenProviderName to SubjectProviderName on AWSStsConfig and
  the runtime AwsStsConfig so it matches token_exchange's name for the
  same concept (which upstream provider's token to pull from
  Identity.UpstreamTokens). Aligns the CRD surface across strategies.
- In the bearer-token fallback, use identity.Token instead of reading
  the outgoing request's Authorization header. The outgoing request is
  the one being constructed for the backend and is not guaranteed to
  carry the caller's header; identity.Token is the original incoming
  bearer captured by the auth middleware. Mirrors token_exchange.
- Log a debug line when falling back to identity.Token.
- Replace sort.Slice with slices.SortStableFunc in buildAwsStsCacheKey
  so mappings sharing a RoleArn but differing in Claim or Matcher keep
  their input order. Otherwise logically identical configs could hash
  to different keys and cause spurious cache misses.
- Include cfg.GetSessionDuration() in the cache key. SessionDuration is
  baked into the cached awsStsContext, so omitting it would let two
  backends differing only in session duration share a cache entry and
  silently use the wrong value.
- Add a success-path case to TestAwsStsStrategy_Validate. NewExchanger
  uses aws.AnonymousCredentials{} so it makes no network calls, and
  the previous test set had only error cases.
- Change authtypes.RoleMapping.Priority from *int32 to *int so it
  actually mirrors awssts.RoleMapping.Priority (the role mapper uses
  math.MaxInt for nil-priority semantics in effectivePriority). Move
  the int32->int conversion to the CRD-to-runtime converter, where it
  belongs.

+++ pkg/vmcp/auth/converters/aws_sts_test.go
@@ -79,7 +79,7 @@ func TestAwsStsConverter(t *testing.T) {
 		assert.Equal(t, "admins", cfg.RoleMappings[0].Claim)
 		assert.Equal(t, "arn:aws:iam::123456789012:role/admin", cfg.RoleMappings[0].RoleArn)
 		require.NotNil(t, cfg.RoleMappings[0].Priority)
-		assert.Equal(t, int32(10), *cfg.RoleMappings[0].Priority)
+		assert.Equal(t, 10, *cfg.RoleMappings[0].Priority)
 		assert.Equal(t, `"devs" in claims["groups"]`, cfg.RoleMappings[1].Matcher)
 		assert.Equal(t, "arn:aws:iam::123456789012:role/dev", cfg.RoleMappings[1].RoleArn)
 	})
